sytle : scope 1 2 3 삭제 모션 통일 - scope 123 파일 명 수정
