+++ src/components/scope12/Scope12SelfInputCaculator.tsx
