+++ src/components/scope3/Scope3CategoryDataInput.tsx
@@ -17,8 +17,11 @@ import {motion, AnimatePresence} from 'framer-motion'
 import {Button} from '@/components/ui/button'
 import {Card, CardContent} from '@/components/ui/card'
 import {Plus, Save} from 'lucide-react'
-import {CalculatorItem} from './CalculatorItem'
-import {scope3CategoryList, Scope3CategoryKey} from '../scopeTotal/CategorySelector'
+import {CalculatorItem} from './Scope3CalculatorItem'
+import {
+  scope3CategoryList,
+  Scope3CategoryKey
+} from '../scopeTotal/Scope123CategorySelector'
 import {SelectorState, Scope3EmissionResponse} from '@/types/scopeTypes'
 import {createScope3Emission, updateScope3Emission} from '@/services/scopeService'
 import {showError, showSuccess, showWarning} from '@/util/toast'
