+++ src/components/scope3/Scope3SelfInputCaculator.tsx
