+++ src/components/scopeTotal/Scope123CategorySelector.tsx
